% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPass-package.r
\docType{package}
\name{getPass-package}
\alias{getPass-package}
\title{getPass}
\description{
A micro-package for reading "passwords", i.e.  reading
user input with masking, so that the input is not displayed as it 
is typed.  Currently we have support for RStudio, the command line
(every OS), and any platform where tcltk is present.
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\references{
Project URL: \url{https://github.com/wrathematics/getPass}
}
\keyword{package}

