% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerobuff.r
\name{zerobuff}
\alias{zerobuff}
\title{Zero Buffers}
\usage{
zerobuff(x)
}
\arguments{
\item{x}{Character; a vector of all character elements.}
}
\value{
\code{NULL} is returned, but all \code{x} will be overwritten by \code{\0}.
}
\description{
Overwirte all characters by \code{\0}.
}
\details{
Overwirte all characters by \code{\0}. This is a very dangeous function
which should be used with causion.
}
\examples{
\dontrun{
x <- c("123", "a", "ab", "abc")
y <- x
zerobuff(x)
print(y)

### zerobuff(letters)  # Don't do this. This may crash entired R.
}

}

