% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPass-package.r
\docType{package}
\name{getPass-package}
\alias{getPass-package}
\title{getPass}
\description{
A micro-package for reading user input with masking, i.e.,
the input is not displayed as it is typed.  Currently the only
supported GUI is a very recent version of RStudio.  Running
from the command line is supported on every major OS.
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\references{
Project URL: \url{https://github.com/wrathematics/getPass}
}
\keyword{package}

