% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPass-package.r
\docType{package}
\name{getPass-package}
\alias{getPass-package}
\title{getPass}
\description{
A micro-package for reading "passwords", i.e.  reading
user input with masking.  The package's only exported function,
\code{getPass()}, behaves similarly to R's own \code{readline()}, only
the user text is not printed as it is typed.  Currently we have support
for RStudio, the command line (every OS), and any platform where tcltk
is present.
}
\references{
Project URL: \url{https://github.com/wrathematics/getPass}
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\keyword{package}
